requirements.txt
streamlit
pandas
pydeck
numpy
altair
esdk-obs-python